Refactor getPayload
